% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_corr.R
\name{fast_corr}
\alias{fast_corr}
\title{Create correlation matrix for columns in dataframe}
\usage{
fast_corr(df, column_names)
}
\arguments{
\item{df}{The data that will be plotted}

\item{column_names}{The vector of column names to be included in correlation matrix}
}
\value{
plot object
}
\description{
The function takes in a dataframe/tibble and a vector of column names
and creates a correlation matrix. 
The correlation matrix can only include numeric columns
}
\examples{
fast_corr(iris, c("Sepal.Length", "Sepal.Width"))
}
