% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_missing_impute.R
\name{fast_missing_impute}
\alias{fast_missing_impute}
\title{Modifies missing data by removing or imputing with mean, median, mode, or multiple.}
\usage{
fast_missing_impute(df, method, cols)
}
\arguments{
\item{df}{The dataframe of interest}

\item{method}{The method of imputation from: {remove, mean, median, mode}}

\item{cols}{The column names with missing data to be modified}
}
\value{
df
}
\description{
The function takes in a dataframe, a list of column names to modify, and
a method of imputation.
The choices of imputation method are either remove (removes all rows with
missing data), mean, median, mode, or multiple imputation.
}
\examples{
fast_missing_impute(df = iris, method = "mean", cols = c("Sepal.Length", "Sepal.Width"))
fast_missing_impute(df = mtcars, method = "median", cols = c("hp", "gear", "carb"))
}
