% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_outliers.R
\name{fast_outlier_id}
\alias{fast_outlier_id}
\title{Analyzes the values of a given column list in a given dataframe, identifies outliers using either the Z-Score algorithm or interquantile range algorithm.
The return is a dataframe containing the following columns: column name, list containing the outlier's index position, percentaje of total counts considered outliers.
Modifies an existing dataframe, with missing values
imputed based on the chosen method.}
\usage{
fast_outlier_id(
  data,
  cols = "All",
  method = "z-score",
  threshold_low_freq = 0.05
)
}
\arguments{
\item{data}{dataframe - Dataframe to be analyzed}

\item{cols}{list -  List containing the columns to be analyzed.}

\item{method}{string - string indicating which method to be used to identify outliers (methods available are: "Z score" or "Interquantile Range")}

\item{threshold_low_freq}{double - Indicates the threshold for evaluating outliers in categorical columns.}
}
\value{
dataframe
}
\description{
Analyzes the values of a given column list in a given dataframe, identifies outliers using either the Z-Score algorithm or interquantile range algorithm.
The return is a dataframe containing the following columns: column name, list containing the outlier's index position, percentaje of total counts considered outliers.
Modifies an existing dataframe, with missing values
imputed based on the chosen method.
}
\examples{
fast_outlier_id(data = iris, cols =  c("Sepal.Length", "Sepal.Width"), method = "z-score")
}
