% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_outliers.R
\name{fast_outliers_id}
\alias{fast_outliers_id}
\title{Analyzes the values of a given column list in a given dataframe, identifies outliers using either the Z-Score algorithm or interquantile range algorithm.
The return is a dataframe containing the following columns: column name, list containing the outlier's index position, percentaje of total counts considered outliers.
Modifies an existing dataframe, with missing values
imputed based on the chosen method.}
\usage{
fast_outliers_id(data, columns, method)
}
\arguments{
\item{data}{dataframe - Dataframe to be analyzed}

\item{columns}{list -  List containing the columns to be analyzed.}

\item{method}{string - string indicating which method to be used to identify outliers (methods available are: "Z score" or "Interquantile Range")}
}
\value{
dataframe
}
\description{
Analyzes the values of a given column list in a given dataframe, identifies outliers using either the Z-Score algorithm or interquantile range algorithm.
The return is a dataframe containing the following columns: column name, list containing the outlier's index position, percentaje of total counts considered outliers.
Modifies an existing dataframe, with missing values
imputed based on the chosen method.
}
\examples{
fast_outliers_identifier(data = iris, method = "Z score", cols =  c("Sepal.Length", "Sepal.Width"))
}
