% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_corr.R
\name{fast_corr}
\alias{fast_corr}
\title{Create correlation matrix for columns in dataframe}
\usage{
fast_corr(df, selected_columns)
}
\arguments{
\item{df}{The data that will be plotted}

\item{selected_columns}{The vector of column names to be included in correlation matrix}
}
\value{
plot object
}
\description{
The function takes in a dataframe/tibble and a vector of column names
and creates a correlation matrix.
The correlation matrix can only include numeric columns
}
\examples{
fast_corr(iris, c('Sepal.Length','Sepal.Width', 'Petal.Length','Petal.Width', 'Species'))
fast_corr(iris, c(1,2,3,4,5))
}
